requirements.txt
aerich
aiogram-dialog==2.0.0b11
aiogram==3.0.0b5
aiohttp
aiosqlite
cachetools
coloredlogs
pyrogram>=2.0.0
redis
tgcrypto
toml
tortoise-orm
gitpython